put all postprocessing of results into PostProcessing class
