+++ ai4water/utils/__init__.py
@@ -1,8 +1,5 @@
 
-from ai4water.utils.utils import make_model, prepare_data
-
-from ai4water.utils.visualizations import PlotResults, murphy_diagram, fdc_plot
-
-from ai4water.utils.taylor_diagram import taylor_plot
-
+from .utils import make_model, prepare_data
+from .visualizations import ProcessResults, murphy_diagram, fdc_plot
+from .taylor_diagram import taylor_plot
 from .easy_mpl import plot, regplot, imshow, bar_chart
