+++ tests/test_postprocessing/test_visualization.py
@@ -11,7 +11,7 @@ import numpy as np
 from ai4water import Model
 from ai4water.datasets import busan_beach
 from ai4water.postprocessing.visualize import Visualize
-from ai4water.utils.visualizations import PlotResults
+from ai4water.utils.visualizations import ProcessResults
 
 
 
@@ -32,37 +32,37 @@ def get_history(keys, add_val=False):
 class TestLossCurve(unittest.TestCase):
 
     def test_plot_loss_1(self):
-        visualizer = PlotResults()
+        visualizer = ProcessResults()
         visualizer.plot_loss(get_history(['loss']))
         visualizer.plot_loss(get_history(['loss'], True))
         return
 
     def test_plot_loss_2(self):
-        visualizer = PlotResults()
+        visualizer = ProcessResults()
         visualizer.plot_loss(get_history(['loss', 'nse']))
         visualizer.plot_loss(get_history(['loss', 'nse'], True))
         return
 
     def test_plot_loss_3(self):
-        visualizer = PlotResults()
+        visualizer = ProcessResults()
         visualizer.plot_loss(get_history(['loss', 'nse', 'r2']))
         visualizer.plot_loss(get_history(['loss', 'nse', 'r2'], True))
         return
 
     def test_plot_loss_4(self):
-        visualizer = PlotResults()
+        visualizer = ProcessResults()
         visualizer.plot_loss(get_history(['loss', 'nse', 'r2', 'kge']))
         visualizer.plot_loss(get_history(['loss', 'nse', 'r2', 'kge'], True))
         return
 
     def test_plot_loss_5(self):
-        visualizer = PlotResults()
+        visualizer = ProcessResults()
         visualizer.plot_loss(get_history(['loss', 'nse', 'r2', 'kge', 'pbias']))
         visualizer.plot_loss(get_history(['loss', 'nse', 'r2', 'kge', 'pbias'], True))
         return
 
     def test_plot_loss_6(self):
-        visualizer = PlotResults()
+        visualizer = ProcessResults()
         visualizer.plot_loss(get_history(['loss', 'nse', 'r2', 'kge', 'pbias', 'bias']))
         visualizer.plot_loss(get_history(['loss', 'nse', 'r2', 'kge', 'pbias', 'bias'], True))
         return
